% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{auspol_theme}
\alias{auspol_theme}
\title{Helper function generate colour palette}
\usage{
auspol_theme(
  p,
  type = c("colour", "fill"),
  extra_colours = NULL,
  extra_values = NULL,
  coord_flip = FALSE,
  palette = NULL,
  legend_pos = FALSE
)
}
\arguments{
\item{p}{ggplot object}

\item{type}{type of scales: c("colour","fill)}

\item{extra_colours}{named vector additional colour (hex) values}

\item{extra_values}{vector with all unique combinations (to assign each a colour)}

\item{coord_flip}{whether to flip coordinate axes}

\item{palette}{additional colour palette for unnamed parties}

\item{legend_pos}{legend position}
}
\value{
ggplot object
}
\description{
Helper function generate colour palette
}
\keyword{plotting}
