% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_import}
\alias{data_import}
\title{Helper function to update/download  data}
\usage{
data_import(file)
}
\arguments{
\item{file}{file to import to the cache}
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
