% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_update}
\alias{data_update}
\title{Helper function to update/download  data}
\usage{
data_update(file = NULL)
}
\arguments{
\item{file}{vectors with file name from repository. By default, downloads all files}
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
