% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{delete_data}
\alias{delete_data}
\title{Helper function to update/download  data}
\usage{
delete_data(file = NULL)
}
\arguments{
\item{file}{to delete - defaults to all of them}
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
