% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_predictions.R
\name{.has_seat_predictions}
\alias{.has_seat_predictions}
\title{Whether this machine has seat predictions to learn from (else: shipped tables)}
\usage{
.has_seat_predictions()
}
\description{
Whether this machine has seat predictions to learn from (else: shipped tables)
}
\keyword{internal}
