% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_forward.R
\name{elections_before}
\alias{elections_before}
\title{Which elections are dated strictly before a target election}
\usage{
elections_before(elections, target_election)
}
\arguments{
\item{elections}{Character vector of election labels.}

\item{target_election}{Label of the election being forecast.}
}
\value{
Logical vector, \code{FALSE} for undated labels.
}
\description{
Vector form of \code{\link[=fit_pairs_for]{fit_pairs_for()}} for tables keyed by an election label.
Under \code{AUSPOL_TIME_FORWARD_FITS=0} it is \code{elections != target_election}.
}
