% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{from_link}
\alias{from_link}
\title{Back-transform from the model scale to shares (percent)}
\usage{
from_link(z, scale = c("logit", "points"))
}
\description{
Back-transform from the model scale to shares (percent)
}
\keyword{internal}
