% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_auspol_bar}
\alias{geom_auspol_bar}
\title{Bar chart, customised for this package.}
\usage{
geom_auspol_bar(include_labels = TRUE, reference_line = NULL, nudge_x = 4, ...)
}
\arguments{
\item{include_labels}{Whether to include numeric labels (TRUE by default)}

\item{reference_line}{Value for reference line. If left empty, no line is added.}

\item{nudge_x}{distance between label and bar}

\item{...}{parameters for {ggplot2} functions.
Label parameters (geom_text_repel()) prefixed with "labels."
Reference line parameters (geom_vline()) prefixed with "ref_line."}
}
\description{
Bar chart, customised for this package.
}
\keyword{plotting}
