% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_auspol_line}
\alias{geom_auspol_line}
\title{Line chart, custommised for this package.}
\usage{
geom_auspol_line(include_labels = TRUE, ...)
}
\arguments{
\item{include_labels}{Whether to include numeric labels (TRUE by default)}

\item{...}{parameters for {ggplot2} geom_segment() (segmnet.prefix), geom_point(), geom_col() and geom_text() (labels. prefix).}
}
\description{
Line chart, custommised for this package.
}
\keyword{plotting}
