% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_auspol_lollipop}
\alias{geom_auspol_lollipop}
\title{Lollipop or bar chart, custommised for this package.}
\usage{
geom_auspol_lollipop(format = "lollipop", include_labels = TRUE, ...)
}
\arguments{
\item{format}{Output format : "lollipop" (default) or "bar".}

\item{include_labels}{Whether to include numeric labels (TRUE by default)}

\item{...}{parameters for {ggplot2} geom_segment() (segmnet.prefix), geom_point(), geom_col() and geom_text() (labels. prefix).}
}
\description{
Lollipop or bar chart, custommised for this package.
}
\keyword{plotting}
