% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_retrieval_house.R
\name{get_auspol_house_data}
\alias{get_auspol_house_data}
\title{Obtain get data from filename}
\usage{
get_auspol_house_data(filename, division, election_year)
}
\arguments{
\item{division}{vector with division names}

\item{election_year}{vector with election years}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
Obtain get data from filename
}
