% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_basic_retrieval.R
\name{get_house_2PP}
\alias{get_house_2PP}
\title{Get 2-party preferred party summary (Coalition vs ALP)}
\usage{
get_house_2PP(
  division = "all",
  year = "all",
  state_abb = "all",
  aggregation = FALSE
)
}
\arguments{
\item{division}{character with division name}

\item{year}{number with election year}

\item{state_abb}{vector with state/territory acronym (e.g. NSW,VIC,QLD,etc.)}

\item{aggregation}{Whether to present division totals (defaults to FALSE)}
}
\value{
dataframe with list of elected MPs
}
\description{
Get 2-party preferred party summary (Coalition vs ALP)
}
\keyword{housegetdata}
