% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_get_data.R
\name{get_house_preferences}
\alias{get_house_preferences}
\alias{get_preferences}
\title{Preferences}
\usage{
get_house_preferences(
  division,
  year,
  polling_places = NULL,
  aggregation = FALSE
)

get_preferences(division, year, polling_places = NULL, aggregation = FALSE)
}
\arguments{
\item{division}{vector with division names}

\item{year}{vector with election years}

\item{polling_places}{list of polling places}

\item{aggregation}{whether to aggregate by division}
}
\value{
dataframe with list of elected MPs
}
\description{
Retrieves preference flow, filterable by election and year. Results can be presented by
polling place - as retrived from the AEC - or aggregated by electoral division.
}
\examples{
\dontrun{
# basic use
get_house_preferences("Wills",2019) |> head(10)
# aggregated version
get_house_preferences("Wills",2019,aggregation = TRUE)
# filtered by polling place
get_house_preferences("Wills",2019, polling_places=c("ABSENT")) |> head(10)

}
}
\keyword{housegetdata}
