% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_get_data.R
\name{get_house_primary_vote}
\alias{get_house_primary_vote}
\title{Primary vote for House Elections}
\usage{
get_house_primary_vote(
  division = NULL,
  year = NULL,
  state_abb = NULL,
  party_abb = NULL,
  aggregation = FALSE,
  polling_places = NULL
)
}
\arguments{
\item{division}{vector with division names}

\item{year}{vector with election years}

\item{state_abb}{vector with state/territory acronym (e.g. NSW,VIC,QLD,etc.)}

\item{party_abb}{vector with party abbreviation (e.g. ALP,LIB,NP,GRN,etc.)}

\item{aggregation}{Whether to present division totals (defaults to FALSE)}

\item{polling_places}{vector with regex for polling places}
}
\value{
sf object with selected polygons
}
\description{
Get primary vote for one or more divisions, for one or more elections.
Data can be filtered by state, political party of polling locations. Results can be presented by
polling station or aggregated by division.
}
\keyword{housegetdata}
