% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_get_data.R
\name{get_house_turnout}
\alias{get_house_turnout}
\title{Election turn out}
\usage{
get_house_turnout(division = NULL, year = NULL)
}
\arguments{
\item{division}{character vector with division names. When left blank, returns all division.}

\item{year}{number vector with election years. When left blank, returns all years.}
}
\value{
data frame turnout numbers
}
\description{
Retrieve election turnout, filterable by division and year
}
\examples{
\dontrun{
# Turnout in Riverina
 get_house_turnout(division="Riverina",yeat)
}
}
\keyword{housegetdata}
