% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_preferences_plot.R
\name{house_2PF_plot}
\alias{house_2PF_plot}
\title{Preferences flow from primary to finalists}
\usage{
house_2PF_plot(
  division,
  year,
  var = "Percent",
  extra_colours = NULL,
  plot_format = "bar",
  include_data = FALSE,
  individualise_IND = TRUE
)
}
\arguments{
\item{division}{Electoral division}

\item{year}{Election year}

\item{var}{Variable to be plotted "Percent" (default) or "Transfer Count"}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{plot_format}{Whether to plot  alluvial chart ("alluvial") or a bar chart ("bar", default).}

\item{include_data}{If set to TRUE, data will be included under <<output_var>>$source_data (defaults to FALSE)}

\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}
}
\value{
preference flow, ggplot2 object
}
\description{
Plot representing flow of preferences from first preferences to candidates in last round.
Can be present as alluvial plot or bar chart, showing votes count or percentages.
}
\examples{
\dontrun{
# Preference flow for Burt, 2022
house_2PF_plot("Burt",2022,plot_format = "alluvial")
# Preference flow for Warringah 2022,
house_2PF_plot("Spence",2013,plot_format = "bar")
}
}
\keyword{houseplots}
