% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_preferences_plot.R
\name{house_2PP_comparison_plot}
\alias{house_2PP_comparison_plot}
\title{Two Party-Preferred Comparison}
\usage{
house_2PP_comparison_plot(
  division = NULL,
  year,
  state = NULL,
  var = "Percentage",
  include_data = TRUE
)
}
\arguments{
\item{division}{Electoral division}

\item{year}{Election year}

\item{state}{If division is left null, use this to select all divisions in one of more states.}

\item{var}{Variable to be plotted "Percentage" (default) or "Votes"}

\item{include_data}{If set to TRUE, data will be included under <<output_var>>$source_data (defaults to FALSE)}
}
\value{
ggplot2 object
}
\description{
Plot with two-party preferred values for one of more divisions, for a given year
Can be present as alluvial plot or bar chart, showing votes count or percentages.
}
\examples{
\dontrun{
# Two party preferred plot for Victoria, 2022
house_2PP_comparison_plot(year=2022,state="VIC")
}
}
\keyword{houseplots}
