% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_preferences_plots.R
\name{house_2PP_historical_plot}
\alias{house_2PP_historical_plot}
\title{Two Party-Preferred Comparison}
\usage{
house_2PP_historical_plot(
  division,
  year = NULL,
  var = "Percentage",
  include_labels = TRUE,
  include_data = TRUE
)
}
\arguments{
\item{division}{Electoral division}

\item{year}{Election year}

\item{var}{Variable to be plotted "Percentage" (default) or "Votes"}

\item{include_labels}{If set to TRUE, the plot will include each value.}

\item{include_data}{If set to TRUE, data will be included under <<output_var>>$source_data (defaults to FALSE)}
}
\value{
ggplot2 object
}
\description{
Plot with two-party preferred values for one of more divisions, for a given year
Can be present as alluvial plot or bar chart, showing votes count or percentages.
}
\examples{
\dontrun{
# Plot historical 2PP for Aston
house_2PP_historical_plot(division="Aston")
}
}
\keyword{houseplots}
