% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_preferences_plots.R
\name{house_preference_flow_plot}
\alias{house_preference_flow_plot}
\title{Plot House of reps preferences flow}
\usage{
house_preference_flow_plot(
  division,
  year,
  var = "Percent",
  exclude_parties = NULL,
  merge_parties = NULL,
  extra_colours = NULL,
  include_data = FALSE
)
}
\arguments{
\item{division}{Electoral division}

\item{year}{Election year}

\item{var}{Variable to be plotted "Percent" (default) or "Preference Count"}

\item{exclude_parties}{vector with party acronyms to exclude from plot}

\item{merge_parties}{list of parties to merge in one line following, the format list(NEWCODE=c(code1,code2,etc.))}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{include_data}{If set to TRUE, output of primary_vote_summary(), will be included under <<output_var>>$source_data (defaults to FALSE)}
}
\value{
preference flow, ggplot2 object
}
\description{
Plot flow of preferences in a division as an alluvial plot.
}
\examples{
\dontrun{
# Preference flow for Wills, 2019
house_preference_flow_plot(division = "Wills",year=2019)
# Preference flow for Warringah 2022,
# excluding two finalists from round 1,
# independent candidate in teal.
house_preference_flow_plot(division = "Warringah",year=2022,
}
}
\keyword{houseplots}
