% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_primary_vote_summary_plots.R
\name{house_primary_comparison_plot}
\alias{house_primary_comparison_plot}
\title{Plot historical changes in primary vote}
\usage{
house_primary_comparison_plot(
  division = NULL,
  year = NULL,
  state = NULL,
  label = "Candidate",
  plotted_variable = "Percentage",
  sort_by_value = TRUE,
  extra_colours = NULL,
  plot_format = "lollipop",
  include_labels = FALSE,
  hor_nudge = 5,
  parties = NULL,
  parties_year = NULL,
  merge_parties = NULL,
  include_others = FALSE,
  include_informal = FALSE,
  individualise_IND = TRUE,
  include_data = TRUE,
  data = NULL
)
}
\arguments{
\item{division}{Name of ONE electoral division}

\item{year}{numeric vector with election years (from 2004), defaults to all.}

\item{state}{Code for one state}

\item{label}{How to label the results, either by Candidate Name ("Name",default), Party Name ("PartyNm") or Party abbreviation ("PartyAb")}

\item{plotted_variable}{Variable to plot, out of "OrdinaryVotes", "Percentage" (default) and Percentage_with_Informal}

\item{sort_by_value}{Whether to sort results by descending order (TRUE by default)}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{plot_format}{Whether to plot  lollipop chart ("lollipop", default) or a bar chart.}

\item{include_labels}{If set to TRUE, the plot will include each value.}

\item{hor_nudge}{if labels are included, separation from chart/dot}

\item{parties}{which parties to include in the summary. All (default), a vector of strings
with the party acronyms (see list_parties()), or a number indicating the top n parties from a certain year.}

\item{parties_year}{If \emph{parties} has is NULL or a number, this indicates if the selection needs to be from
a certain year (.e.g only select the historical data for the three top parties in 2012)}

\item{merge_parties}{list of parties to merge in one line following, the format list(NEWCODE=c(code1,code2,etc.))}

\item{include_others}{Boolean used along \emph{parties} to included the remaining votes in one "Other" category.}

\item{include_informal}{Boolean to add informal votes in addition to the party selection.
Informal votes will be included if no parties are selected, or the top n parties are selected,
and it happens to be in the top n - even if this flag is set to false.}

\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}

\item{include_data}{If set to TRUE, output of house_primary_vote_summary(), will be included under <<output_var>>$source_data (defaults to FALSE)}

\item{data}{Alternative, instead of providing a parameters, it is possible to provide the data frame with the data
to plot, following the format from the output of  house_primary_vote_summary().}
}
\value{
ggplot2 object
}
\description{
Line chart with historial changes for a division, group of candidates in a party,
selected parties, etc.
}
\examples{
\dontrun{
# Compare primary voting in Kooyong in 2022,  lollipop chart (default)
house_primary_comparison_plot(division = "Kooyong",
                              year=2022,
                              individualise_IND = TRUE)
# Liberal Primary vote in Tasmania in 2022, bar chart
house_primary_comparison_plot(state="TAS",
                              year=2022,
                              parties=c("LP"),
                              plot_format = "bar")

}
}
\keyword{houseplots}
