% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_primary_vote_summary_plots.R
\name{house_primary_historic_plot}
\alias{house_primary_historic_plot}
\title{Plot primary vote history}
\usage{
house_primary_historic_plot(
  division = NULL,
  plotted_variable = "Percentage",
  parties = NULL,
  parties_year = NULL,
  merge_parties = NULL,
  include_others = FALSE,
  include_informal = FALSE,
  individualise_IND = FALSE,
  extra_colours = NULL,
  include_labels = FALSE,
  year = NULL,
  include_data = FALSE,
  data = NULL
)
}
\arguments{
\item{division}{named vector with division names}

\item{plotted_variable}{Variable to plot, out of "OrdinaryVotes", "Percentage" (default) and Percentage_with_Informal}

\item{parties}{which parties to include in the summary. All (default), a vector of strings
with the party acronyms (see list_parties()), or a number indicating the top n parties from a certain year.}

\item{parties_year}{If \emph{parties} has is NULL or a number, this indicates if the selection needs to be from
a certain year (.e.g only select the historical data for the three top parties in 2012).}

\item{merge_parties}{list of parties to merge in one line following, the format list(NEWCODE=c(code1,code2,etc.))}

\item{include_others}{Boolean used along \emph{parties} to included the remaining votes in one "Other" category.}

\item{include_informal}{Boolean to add informal votes in addition to the party selection.
Informal votes will be included if no parties are selected, or the top n parties are selected,
and it happens to be in the top n - even if this flag is set to false.}

\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{include_labels}{If set to TRUE, the plot will include each value.}

\item{year}{numeric vector with election years (from 2004), defaults to all.}

\item{include_data}{If set to TRUE, output of house_primary_vote_summary(), will be included under <<output_var>>$source_data (defaults to FALSE)}

\item{data}{Alternative, instead of providing a parameters, it is possible to provide the data frame with the data
to plot, folowing the format from the output of  house_primary_vote_summary().}
}
\value{
ggplot2 object
}
\description{
Plot historical primary vote results for a division or group of divisions, being able to select
and aggregate political parties. Can plot either percentages or absolute number of ordinary votes.
}
\examples{
\dontrun{
#

}
}
\keyword{houseplots}
