% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_results_plots.R
\name{house_results_historic}
\alias{house_results_historic}
\title{Historic Resuls}
\usage{
house_results_historic(
  individualise_IND = FALSE,
  merge_parties = NULL,
  parties = NULL,
  include_others = FALSE,
  include_labels = TRUE,
  extra_colours = NULL,
  include_data = FALSE
)
}
\arguments{
\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}

\item{merge_parties}{list of parties to merge in one line following, the format list(NEWCODE=c(code1,code2,etc.))}

\item{parties}{List of political party abbreviations to filter on. If merge_parties is used, those names can be included too.}

\item{include_others}{Boolean used along \emph{parties} to included the remaining votes in one "Other" category.}

\item{include_labels}{If set to TRUE, the plot will include each value.}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{include_data}{If set to TRUE, data will be included under <<output_var>>$source_data (defaults to FALSE)}
}
\value{
preference flow, ggplot2 object
}
\description{
Plot seats by party across time. Parties can be filtered and grouped by coalitions.
}
\examples{
\dontrun{
# Historic results, focusing showing tally for Coalition, ALP, Greens - others merged together
house_results_historic(merge_parties = list(COAL=c("CLP","LP","LNP","NP")),
                       parties =c("COAL","ALP","GRN"),
                       include_other=TRUE))
}
}
\keyword{houseplots}
