% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{import_data}
\alias{import_data}
\title{Helper function to update/download  data}
\usage{
import_data(file)
}
\arguments{
\item{file}{file to import to the cache}
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
