% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_divisions}
\alias{list_divisions}
\title{Get list of divisions}
\usage{
list_divisions(filters = NULL)
}
\arguments{
\item{filters}{\emph{(optional)} list() with filters in the form list(Column="Value")}
}
\value{
data frame with lists of divisions
}
\description{
get list of all the Australian Federal electoral divisions, being able to
filter by any attribute. Covers all divisions from the 2004 Election.
}
\examples{
\dontrun{
# Get list of all divisions
list_divisions()

#Get list containing only Wills and Melbourne
list_divisions(filters=list(DivisionNm=c("Wills","Melbourne")))
 }
}
\keyword{housegetdata}
\keyword{lists}
