% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_polling_stations}
\alias{list_polling_stations}
\title{List all polling stations}
\usage{
list_polling_stations()
}
\value{
data frame with lists of polling stations
}
\description{
List all polling stations
}
\examples{
\dontrun{
# Get list of all registered parties
list_parties()
 }
}
\keyword{lists}
