% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_years}
\alias{list_years}
\title{Get election years.}
\usage{
list_years()
}
\value{
vector with years
}
\description{
Very simple function listing the election years included in this package.
}
\examples{
\dontrun{
# Get list of all divisions
list_years()
 }
}
\keyword{lists}
