% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_auspol}
\alias{load_auspol}
\title{Helper function to download  data}
\usage{
load_auspol(
  auspol_file,
  refresh = getOption("auspol_refresh", FALSE),
  auspol_type = NULL,
  class = getOption("auspol_class", "sf"),
  progress_bar = TRUE,
  keep_zipped_shapefile = FALSE
)
}
\arguments{
\item{auspol_file}{name of the file to download.}

\item{refresh}{Whether to re-download shapefiles if cached. Defaults to value of the global
option "auspol_refresh" if that option is, and FALSE if not. This will override the behavior
set in "auspol_refresh" option if a value (TRUE or FALSE) is provided.}

\item{auspol_type}{Added as an attribute to return object (used internally).}

\item{class}{Class of return object. Must be one of "sf" (the default) or "sp".}

\item{progress_bar}{If set to FALSE, do not display download progress bar
(helpful for R Markdown documents). Defaults to TRUE.}

\item{keep_zipped_shapefile}{If set to TRUE, do not delete zipped shapefile
(stored in temporary directory or AUSPOL_CACHE_DIR depending on the configuration of
global option "auspol_use_cache"). Defaults to FALSE.}
}
\value{
dataframe
}
\description{
Helper function to download  data
}
