% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_polls.R
\name{load_election_cycles}
\alias{load_election_cycles}
\title{Load election cycle boundaries}
\usage{
load_election_cycles()
}
\value{
data.table: \code{year}, \code{region}, \code{start}, \code{end} (election day).
}
\description{
Load election cycle boundaries
}
