% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{load_federal_situation}
\alias{load_federal_situation}
\title{Which party held federal government at the time of each state election}
\usage{
load_federal_situation()
}
\value{
data.table: \code{year}, \code{region}, \code{fed_govt}, \code{fed_opp}.
}
\description{
Which party held federal government at the time of each state election
}
