% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{load_incumbency}
\alias{load_incumbency}
\title{Who was in government at each election, and for how long}
\usage{
load_incumbency()
}
\value{
data.table: \code{year}, \code{region}, \code{incumbent}, \code{opposition}, \code{years}.
}
\description{
Who was in government at each election, and for how long
}
