% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{load_polled_elections}
\alias{load_polled_elections}
\title{Elections for which voting-intention polling exists}
\usage{
load_polled_elections()
}
\value{
data.table: \code{year}, \code{region}.
}
\description{
Elections for which voting-intention polling exists
}
