% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{manage_colours}
\alias{manage_colours}
\title{Helper function generate colour palette}
\usage{
manage_colours(extra_colours = NULL, extra_values = NULL, palette = NULL)
}
\arguments{
\item{extra_colours}{named vector additional colour (hex) values}

\item{extra_values}{vector with all unique combinations (to assign each a colour)}

\item{palette}{palette to replace brewer.pal}
}
\value{
named vector with colours
}
\description{
Helper function generate colour palette
}
\keyword{plotting}
