% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{party_colours}
\alias{party_colours}
\title{Named vector with common party colours, with option to add custom/additional values}
\usage{
party_colours(extra = NULL)
}
\arguments{
\item{extra}{named vector additional colour (hex) values}
}
\value{
named vector
}
\description{
Named vector with common party colours, with option to add custom/additional values
}
\keyword{plotting}
