% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{pkg_root}
\alias{pkg_root}
\title{Repository root (works from package load or scripts run at repo root)}
\usage{
pkg_root()
}
\description{
Repository root (works from package load or scripts run at repo root)
}
\keyword{internal}
