% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_vote_summary_plots.R
\name{plot_preference_flow}
\alias{plot_preference_flow}
\title{Plot preference flow, for a division for a given year}
\usage{
plot_preference_flow(
  division,
  year,
  var = "Percent",
  exclude_parties = NULL,
  extra_colours = NULL,
  include_data = FALSE
)
}
\arguments{
\item{division}{Electoral division}

\item{year}{Election year}

\item{var}{Variable to be plotted "Percent" (default) or "Preference Count"}

\item{exclude_parties}{vector with party acronyms to exclude from plot}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{include_data}{If set to TRUE, output of primary_vote_summary(), will be included under <<output_var>>$source_data (defaults to FALSE)}
}
\value{
preference flow, ggplot2 object
}
\description{
Plot preference flow, for a division for a given year
}
\keyword{houseplots}
