% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_vote_summary_plots.R
\name{plot_primary_comparison}
\alias{plot_primary_comparison}
\title{Plot historical changes in primary vote in a electoral division (line chart)}
\usage{
plot_primary_comparison(
  division = NULL,
  year = NULL,
  state = NULL,
  label = "Candidate",
  plotted_variable = "Percentage",
  sort_by_value = TRUE,
  extra_colours = NULL,
  plot_format = "lollipop",
  include_labels = FALSE,
  nudge_x = 5,
  parties = NULL,
  parties_year = NULL,
  include_others = FALSE,
  include_informal = FALSE,
  individualise_IND = FALSE,
  include_data = TRUE,
  data = NULL
)
}
\arguments{
\item{division}{Name of ONE electoral division}

\item{year}{numeric vector with election years (from 2004), defaults to all.}

\item{state}{Code for one state}

\item{label}{How to label the results, either by Candidate Name ("Name",default), Party Name ("PartyNm") or Party abbreviation ("PartyAb")}

\item{plotted_variable}{Variable to plot, out of "OrdinaryVotes", "Percentage" (default) and Percentage_with_Informal}

\item{sort_by_value}{Whether to sort results by descending order (TRUE by default)}

\item{extra_colours}{manual mapping of colours for each party, as a named vector.}

\item{plot_format}{Whether to plot  lollipop chart ("lollipop", default) or a bar chart.}

\item{include_labels}{If set to TRUE, the plot will include each value.}

\item{nudge_x}{if labels are included, separation from chart/dot}

\item{parties}{which parties to include in the summary. All (default), a vector of strings
with the party acronyms (see list_parties()), or a number indicating the top n parties from a certain year.}

\item{parties_year}{If \emph{parties} has is NULL or a number, this indicates if the selection needs to be from
a certain year (.e.g only select the historical data for the three top parties in 2012)}

\item{include_others}{Boolean used along \emph{parties} to included the remaining votes in one "Other" category.}

\item{include_informal}{Boolean to add informal votes in addition to the party selection.
Informal votes will be included if no parties are selected, or the top n parties are selected,
and it happens to be in the top n - even if this flag is set to false.}

\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}

\item{include_data}{If set to TRUE, output of primary_vote_summary(), will be included under <<output_var>>$source_data (defaults to FALSE)}

\item{data}{Alternative, instead of providing a parameters, it is possible to provide the data frame with the data
to plot, following the format from the output of  primary_vote_summary().}
}
\value{
ggplot2 object
}
\description{
Plot historical changes in primary vote in a electoral division (line chart)
}
\keyword{houseplots}
