% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_basic_retrieval.R
\name{preference_flow_data}
\alias{preference_flow_data}
\title{Get preference flow (rounds) for en electorate on a given election}
\usage{
preference_flow_data(
  division,
  election_year,
  individualise_IND = TRUE,
  parties_exclude = NULL,
  rounds_exclude = 0
)
}
\arguments{
\item{division}{division}

\item{election_year}{election year}
}
\value{
list with data frames with results for each round
}
\description{
Get preference flow (rounds) for en electorate on a given election
}
\keyword{housegetdata}
