% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_vote_summary_data.R
\name{primary_vote_summary}
\alias{primary_vote_summary}
\title{Helper function to download  data}
\usage{
primary_vote_summary(
  division = NULL,
  state = NULL,
  year = NULL,
  parties = NULL,
  parties_year = NULL,
  include_others = FALSE,
  include_informal = FALSE,
  include_names = TRUE,
  individualise_IND = FALSE,
  wide_format = NULL
)
}
\arguments{
\item{division}{vector with names of electoral divisions (e.g. "Banks", "Wills","Indi")}

\item{state}{if divisions are not provide, provide a vector with state initials e.g. c("NT","TAS")}

\item{year}{numeric vector with election years (from 2004), defaults to all.}

\item{parties}{which parties to include in the summary. All (default), a vector of strings
with the party acronyms (see list_parties()), or a number indicating the top n parties from a certain year.}

\item{parties_year}{If \emph{parties} has is NULL or a number, this indicates if the selection needs to be from
a certain year (.e.g only select the historical data for the three top parties in 2012)}

\item{include_others}{Boolean used along \emph{parties} to included the remaining votes in one "Other" category.}

\item{include_informal}{Boolean to add informal votes in addition to the party selection.
Informal votes will be included if no parties are selected, or the top n parties are selected,
and it happens to be in the top n - even if this flag is set to false.}

\item{include_names}{whether to include the candidates name and surname in the extract (TRUE by default).}

\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}

\item{wide_format}{Whether to present the result in long format, like the AEC's source, or a year-by-year summary.
Options include NULL (no summarisation, default), "OrdinaryVotes" (absolute numbers), "Percentage_with_Informal"
and  "Percentage" (which is the percentage counted on elections).}
}
\value{
dataframe
}
\description{
Helper function to download  data
}
\keyword{houseconvenience}
