% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{sd_from_link}
\alias{sd_from_link}
\title{Convert an sd on the model scale back to percentage points (delta method)}
\usage{
sd_from_link(sd_link, p, scale = c("logit", "points"))
}
\description{
Convert an sd on the model scale back to percentage points (delta method)
}
\keyword{internal}
