% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{sd_to_link}
\alias{sd_to_link}
\title{Convert an sd in percentage points to the model scale (delta method)}
\usage{
sd_to_link(sd_pts, p, scale = c("logit", "points"))
}
\arguments{
\item{sd_pts}{Standard deviation in percentage points.}

\item{p}{Share (percent) at which to linearise.}
}
\description{
Convert an sd in percentage points to the model scale (delta method)
}
\keyword{internal}
