% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{SHARE_CLAMP}
\alias{SHARE_CLAMP}
\title{Smallest and largest share (percent) representable on the logit scale}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
SHARE_CLAMP
}
\description{
Smallest and largest share (percent) representable on the logit scale
}
\keyword{internal}
