% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{to_link}
\alias{to_link}
\title{Transform poll shares (percent) to the model scale}
\usage{
to_link(y, scale = c("logit", "points"))
}
\arguments{
\item{y}{Shares in percent.}

\item{scale}{"logit" or "points".}
}
\value{
List: \code{z} (model scale), \code{n_clamped}.
}
\description{
Transform poll shares (percent) to the model scale
}
\keyword{internal}
