% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_basic_retrieval.R
\name{get_house_2PF}
\alias{get_house_2PF}
\title{Get 2-party preferred flow (as calculated by the ABS), for an division on a given election}
\usage{
get_house_2PF(division, year, aggregation = FALSE)
}
\arguments{
\item{division}{character with division name}

\item{year}{number with election year}

\item{aggregation}{Whether to present division totals (defaults to FALSE)}
}
\value{
dataframe with list of elected MPs
}
\description{
Get 2-party preferred flow (as calculated by the ABS), for an division on a given election
}
\keyword{housegetdata}
