% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_get_data.R
\name{get_house_MPs}
\alias{get_house_MPs}
\alias{get_MPs}
\title{Elected MPs}
\usage{
get_house_MPs(division = NULL, year = NULL)

get_MPs(division = NULL, year = NULL)
}
\arguments{
\item{division}{character vector with division names. When left blank, returns all division.}

\item{year}{number vector with election years. When left blank, returns all years.}
}
\value{
data frame with list of elected MPs
}
\description{
Retrieve list of elected MPs, filterable by division and year
}
\examples{
\dontrun{
# Elected MPs in Melbourne and Cooper, 2019 and 2022
get_house_MPs(division = c("Melbourne","Cooper"),
              year = c(2019,2022))
}
}
\keyword{housegetdata}
