% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_retrieval_house.R
\name{get_house_MPs}
\alias{get_house_MPs}
\title{Reetrieve elected MPs, for groups of divisions, for given year}
\usage{
get_house_MPs(division = "all", election_year = "all")
}
\arguments{
\item{division}{vector with division names}

\item{election_year}{vector with election years}
}
\value{
dataframe with list of elected MPs
}
\description{
Reetrieve elected MPs, for groups of divisions, for given year
}
\keyword{house_get_data}
