% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_get_data.R
\name{get_house_preferences}
\alias{get_house_preferences}
\alias{get_preferences}
\title{Get list of with turnout all elected MPs, for groups of divisions, for given year}
\usage{
get_house_preferences(
  division,
  year,
  polling_places = NULL,
  aggregation = TRUE
)

get_preferences(division, year, polling_places = NULL, aggregation = TRUE)
}
\arguments{
\item{division}{vector with division names}

\item{year}{vector with election years}

\item{polling_places}{list of polling places}

\item{aggregation}{whether to aggregate by division}
}
\value{
dataframe with list of elected MPs
}
\description{
Get list of with turnout all elected MPs, for groups of divisions, for given year
}
\examples{
\dontrun{
# basic use
get_house_preferences("Wills",2019)
# disaggregated version
get_house_preferences("Wills",2019,aggregation = FALSE) |> head(10)
# filtered by polling place
get_house_preferences("Wills",2019, polling_places=c("ABSENT"),aggregation = FALSE) |> head(10)
}
}
\keyword{housegetdata}
