% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_get_data.R
\name{get_house_primary_vote}
\alias{get_house_primary_vote}
\title{Primary vote for House Elections}
\usage{
get_house_primary_vote(
  division = NULL,
  year = NULL,
  state_abb = NULL,
  party_abb = NULL,
  aggregation = FALSE,
  polling_places = NULL
)
}
\arguments{
\item{division}{vector with division names}

\item{year}{vector with election years}

\item{state_abb}{vector with state/territory acronym (e.g. NSW,VIC,QLD,etc.)}

\item{party_abb}{vector with party abbreviation (e.g. ALP,LIB,NP,GRN,etc.)}

\item{aggregation}{Whether to present division totals (defaults to FALSE)}

\item{polling_places}{vector with regex for polling places}
}
\value{
sf object with selected polygons
}
\description{
Get primary vote for one or more divisions, for one or more elections.
Data can be filtered by state, political party of polling locations. Results can be presented by
polling station or aggregated by division.
}
\examples{
\dontrun{
# Primary vote in Brisbane, 2022 election
get_house_primary_vote(division="Brisbane",year=2022)
# Primary vote in Perth and Brisbane in 2019 and 2022 (aggregated)
get_house_primary_vote(division=c("Brisbane","Perth"),year=c(2019,2022),aggregation = TRUE)
# Primary vote for Greens candidates in Tasmania and the Northern Territory, 2019
get_house_primary_vote(state=c("TAS","NT"),year=2019,aggregation = TRUE, party_abb=c("GRN"))
}
}
\keyword{housegetdata}
