% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_basic_retrieval.R
\name{get_house_primary_vote}
\alias{get_house_primary_vote}
\title{Obtain primary vote per candidate for a given set of divisions, for a set of years}
\usage{
get_house_primary_vote(
  division = "all",
  year = "all",
  state_abb = "all",
  party_abb = "all",
  aggregation = FALSE,
  polling_places = NULL
)
}
\arguments{
\item{division}{vector with division names}

\item{year}{vector with election years}

\item{state_abb}{vector with state/territory acronym (e.g. NSW,VIC,QLD,etc.)}

\item{party_abb}{vector with party abbreviation (e.g. ALP,LIB,NP,GRN,etc.)}

\item{aggregation}{Whether to present division totals (defaults to FALSE)}

\item{polling_places}{vector with polling places}
}
\value{
sf object with selected polygons
}
\description{
Obtain primary vote per candidate for a given set of divisions, for a set of years
}
\keyword{housegetdata}
