% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_retrieval_house.R
\name{get_house_primary_vote}
\alias{get_house_primary_vote}
\title{Obtain primary vote per candidate for a given set of divisions, for a set of years}
\usage{
get_house_primary_vote(
  division = "all",
  election_year = "all",
  state_abb = "all",
  party_abb = "all",
  aggregation = FALSE
)
}
\arguments{
\item{division}{vector with division names}

\item{election_year}{vector with election years}

\item{state_abb}{vector with state/territory acronym (e.g. NSW,VIC,QLD,etc.)}

\item{party_abb}{vector with party abbreviation (e.g. ALP,LIB,NP,GRN,etc.)}

\item{aggregation}{Wether to present division totals (defaults to FALSE)}
}
\value{
sf object with selected polygons
}
\description{
Obtain primary vote per candidate for a given set of divisions, for a set of years
}
\keyword{#}
\keyword{#'}
\keyword{2019}
\keyword{2022}
\keyword{@examples}
\keyword{Candidates}
\keyword{Cooper}
\keyword{Get}
\keyword{Green}
\keyword{Victoria,}
\keyword{Wills}
\keyword{\dontrun{}
\keyword{across}
\keyword{all}
\keyword{and}
\keyword{election}
\keyword{for}
\keyword{get_house_primary_vote(c("Wills","Cooper),2022)}
\keyword{get_house_primary_vote(election_year=2019,state_abb="VIC",party_abb="GRN")}
\keyword{house_get_data}
\keyword{in}
\keyword{list}
\keyword{of}
\keyword{primary}
\keyword{vote}
\keyword{votes}
\keyword{}}
