% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_basic_retrieval.R
\name{get_house_turnout}
\alias{get_house_turnout}
\title{Get list of with turnout all elected MPs, for groups of divisions, for given year}
\usage{
get_house_turnout(division = "all", year = "all")
}
\arguments{
\item{division}{character with division name}

\item{year}{number with election year}
}
\value{
dataframe with list of elected MPs
}
\description{
Get list of with turnout all elected MPs, for groups of divisions, for given year
}
\keyword{housegetdata}
