% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_divisions}
\alias{list_divisions}
\title{List all divisions}
\usage{
list_divisions()
}
\value{
data frame with lists of divisions
}
\description{
List all divisions
}
\examples{
\dontrun{
# Get list of all divisions
list_divisions()
 }
}
\keyword{lists}
