% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_parties}
\alias{list_parties}
\title{List all political parties.}
\usage{
list_parties(filters = NULL, party_regex = NULL)
}
\arguments{
\item{filters}{(optional) list() with filters in the form list(Column="Value").}

\item{party_regex}{additional filter for party names, taking a regular expression.}
}
\value{
data frame with lists of divisions
}
\description{
Lists all political parties that have participated from the 2004 Election onwards.
Parties are presented as recorded by the AEC. List can be filtered by party names matching a regular expression.
}
\examples{
\dontrun{
# Get list of all registered political parties
list_parties()
#
# Get list of all parties whose name start with "Australia"
list_parties(party_regex="^Australia")
 }
}
\keyword{lists}
