% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_parties}
\alias{list_parties}
\title{List all political parties participating in a election year, for a state, or matching a pattern in their full name}
\usage{
list_parties()
}
\value{
data frame with lists of divisions
}
\description{
List all political parties participating in a election year, for a state, or matching a pattern in their full name
}
\examples{
\dontrun{
# Get list of all divisions in 2013
list_parties()
 }
}
\keyword{lists}
