% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_polling_places}
\alias{list_polling_places}
\title{List all polling stations}
\usage{
list_polling_places(filters = NULL)
}
\arguments{
\item{filters}{(optional) list() with filters in the form list(Column="Value")}
}
\value{
data frame with lists of polling stations
}
\description{
Retrieve list of all polling station that been used from 2044 onwards.
Names as recorded by the AEC. List can be filtered by state, division names and
regular expressions matching their names.
}
\examples{
\dontrun{
# Get list of all registered parties
list_parties()
# Get list of polling places in the division of Hasluck
list_parties(list)

 }
}
\keyword{lists}
