% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_plots.R
\name{party_colours}
\alias{party_colours}
\title{Named vector with common parry colours, with option to add custom/additonal values}
\usage{
party_colours(extra = NULL)
}
\arguments{
\item{extra}{named vector additional colour (hex) values}
}
\value{
named vector
}
\description{
Named vector with common parry colours, with option to add custom/additonal values
}
\keyword{helpers}
