% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_vote_summary_data.R
\name{preference_flow_data}
\alias{preference_flow_data}
\title{Get preference flow (rounds) for en electorate on a given election}
\usage{
preference_flow_data(
  division,
  election_year,
  individualise_IND = TRUE,
  exclude_parties = NULL,
  exclude_rounds = 0
)
}
\arguments{
\item{division}{division}

\item{election_year}{election year}

\item{individualise_IND}{If set to TRUE, party abbreviations for each independent candidate will be changed
from "IND" to "IND-<<candidate's surname>>", effectively separating them in party aggregations.}

\item{exclude_parties}{vector with party acronyms to exclude from plot}

\item{exclude_rounds}{If parties are excluded, include vector indicating from which rounds should them be excluded}
}
\value{
list with data frames with results for each round
}
\description{
Get preference flow (rounds) for en electorate on a given election
}
\keyword{housegetdata}
