% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{trend_anchor}
\alias{trend_anchor}
\title{Day-0 anchor rule (previous election result, or loosely the first poll)}
\usage{
trend_anchor(prep, prior_result)
}
\description{
Day-0 anchor rule (previous election result, or loosely the first poll)
}
\keyword{internal}
